refactor: 금액 계산중 표현범위를 벗어나는 부분을 Big(Number)로 변경한다

+++ src/main/java/lottogame/console/ConsoleInputer.java
@@ -11,10 +11,10 @@ import lottogame.controller.spi.Inputer;
 public class ConsoleInputer implements Inputer {
 
     @Override
-    public int inputMoney() {
+    public long inputMoney() {
         LocalScanner localScanner = LocalScanner.getInstance();
         System.out.println("구입금액을 입력해 주세요.");
-        return localScanner.nextInt();
+        return localScanner.nextLong();
     }
 
     @Override
@@ -51,4 +51,5 @@ public class ConsoleInputer implements Inputer {
         System.out.println("보너스 볼을 입력해 주세요.");
         return LocalScanner.getInstance().nextInt();
     }
+
 }
