+++ src/main/java/lottogame/domain/LottoPurchaseManager.java
@@ -15,10 +15,10 @@ public class LottoPurchaseManager {
         this.numberGenerator = numberGenerator;
     }
 
-    public List<LottoTicketDto> purchase(int money, List<LottoTicketDto> lottoTicketDtos) {
+    public List<LottoTicketDto> purchase(long money, List<LottoTicketDto> lottoTicketDtos) {
         assertMoney(money);
         assertPassiveLottoTickets(money, lottoTicketDtos);
-        money -= lottoTicketDtos.size() * LottoTicket.PURCHASABLE_UNIT;
+        money -= (long) lottoTicketDtos.size() * LottoTicket.PURCHASABLE_UNIT;
         List<LottoTicket> lottoTickets = lottoTicketDtos.stream()
             .map(lottoTicketDto -> new LottoTicket(lottoTicketDto.getNumbers()))
             .collect(Collectors.toList());
@@ -26,7 +26,7 @@ public class LottoPurchaseManager {
         return toLottoTicketDtos(lottoTickets);
     }
 
-    private void assertMoney(int money) {
+    private void assertMoney(long money) {
         if (money % LottoTicket.PURCHASABLE_UNIT != 0) {
             throw new IllegalArgumentException(
                 String.format("money는 \"%d\"원으로 나누어 떨어져야 합니다 money: \"%d\"", LottoTicket.PURCHASABLE_UNIT,
@@ -34,15 +34,15 @@ public class LottoPurchaseManager {
         }
     }
 
-    private void assertPassiveLottoTickets(int money, List<LottoTicketDto> lottoTicketDtos) {
-        int lottoBuyableCount = money / LottoTicket.PURCHASABLE_UNIT;
+    private void assertPassiveLottoTickets(long money, List<LottoTicketDto> lottoTicketDtos) {
+        long lottoBuyableCount = money / LottoTicket.PURCHASABLE_UNIT;
         if (lottoBuyableCount < lottoTicketDtos.size()) {
             throw new IllegalArgumentException(String.format("구매가능한 로또 숫자 \"%d\" 보다 선택한 로또의 숫자가 더 많습니다.",
                 lottoBuyableCount));
         }
     }
 
-    private List<LottoTicket> createLottoTickets(int money) {
+    private List<LottoTicket> createLottoTickets(long money) {
         List<LottoTicket> lottoTicketList = new ArrayList<>();
         for (int count = 0; count < money / LottoTicket.PURCHASABLE_UNIT; count++) {
             lottoTicketList.add(new LottoTicket(numberGenerator));
